fix pylint issues
