+++ scripts/lib/CIME/build.py
@@ -1,10 +1,11 @@
 """
 functions for building CIME models
 """
+import glob, shutil, time, threading, gzip, subprocess
 from CIME.XML.standard_module_setup  import *
 from CIME.utils                 import get_model, analyze_build_log, stringify_bool, run_and_log_case_status, get_timestamp, run_sub_or_cmd, run_cmd, get_batch_script_for_job
 from CIME.provenance            import save_build_provenance
-import glob, shutil, time, threading, gzip, subprocess
+from CIME.locked_files          import lock_file, unlock_file
 
 logger = logging.getLogger(__name__)
